show default rate, round for 2 signs after point

+++ app/src/main/java/ru/bagrusss/revolutdemo/net/models/CurrenciesResponse.kt
@@ -2,6 +2,7 @@ package ru.bagrusss.revolutdemo.net.models
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
+import java.math.BigDecimal
 
 /**
  * Created by bagrusss on 12.08.2019
@@ -10,5 +11,5 @@ import com.squareup.moshi.JsonClass
 class RatesResponse(
     @Json(name = "date") val date: String,
     @Json(name = "base") val base: String,
-    @Json(name = "rates") val rates: Map<String, Double>
+    @Json(name = "rates") val rates: Map<String, Float>
 )
